Refactor evaluation imports, remove unused files, add feature and store creation boilerplate
