+++ src/app/evaluations/page.tsx
@@ -1,4 +1,4 @@
-import Main from "./Main";
+import Main from "../../features/evaluations/presentation/components/Main";
 
 export default function Page() {
   return <Main />;
