+++ src/features/evaluations/common/evaluation-store.ts
@@ -1,0 +1,20 @@
+import { create } from "zustand";
+import { persist } from "zustand/middleware";
+
+type EvaluationCreationState = {
+  isCreating: boolean;
+  setIsCreating: (isCreating: boolean) => void;
+};
+
+export const useEvaluationCreationStore = create<EvaluationCreationState>()(
+  persist(
+    (set, get) => ({
+      isCreating: false,
+      setIsCreating: () => set((state) => ({ isCreating: !state.isCreating })),
+    }),
+    {
+      name: "evaluation-creation",
+      skipHydration: true,
+    }
+  )
+);
